[opt, fix] Move some imports out of index files
Node runtime for jest was not playing nicely for some runtime-only
exports from index files. Move them out so they are only accessible
through static analysis and browser runtime.
